[Move] Profile関連のvueファイルをsharedからcomponentに移動
