+++ app/javascript/packs/pages/mypage.vue
@@ -6,7 +6,7 @@
 </template>
 <script>
 import Axios from "axios"
-import Profile from "../components/shared/Profile"
+import Profile from "../components/Profile"
 import TagsTab from "../components/TagsTab"
 
 export default {
